feat: 重构环境和工具链设置视图，优化样式和功能，添加工具链配置页面
