+++ src/vue/setting/data.ts
@@ -6,72 +6,6 @@ export let extensionInfo = ref<any>({
     version: "0.4.11"
 });
 
-export const analysisInfo = ref<any>({
-    mapFile: "rt-thread.map",
-    mapFileUrl: "",
-    analysisTableList: [
-        {
-            title: "全部符号",
-            children: [
-                {
-                    name: "name01",
-                    path: "0x00",
-                    size: "12",
-                },
-                {
-                    name: "name02",
-                    path: "0x20",
-                    size: "16",
-                },
-            ]
-        },
-        {
-            title: "按文件",
-            children: [
-                {
-                    name: "file01",
-                    path: "0x00",
-                    size: "12",
-                },
-                {
-                    name: "file02",
-                    path: "0x20",
-                    size: "16",
-                },
-            ]
-        },
-        {
-            title: "按Groups",
-            children: [
-                {
-                    name: "Groups01",
-                    path: "0x00",
-                    size: "12",
-                },
-                {
-                    name: "Groups02",
-                    path: "0x20",
-                    size: "16",
-                },
-            ]
-        }
-    ],
-    analysisTitleList: [
-        {
-            title: "名称",
-            field: "name",
-        },
-        {
-            title: "地址",
-            field: "path",
-        },
-        {
-            title: "大小",
-            field: "size",
-        },
-    ]
-});
-
 export const envInfo = ref<any>({
     progressNum: 0,
     selectRow: null,
