+++ src/vue/setting/view/toolchains/index.less
@@ -1,0 +1,81 @@
+/* SDK管理器页面样式 */
+.content {
+    padding: 20px;
+    height: 100%;
+    box-sizing: border-box;
+}
+
+.body-box {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.sdk-title {
+    font-size: 18px;
+    font-weight: bold;
+    color: #333;
+    margin-bottom: 20px;
+    padding-bottom: 10px;
+    border-bottom: 2px solid #409eff;
+}
+
+.table-box {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+}
+
+.el-table {
+    flex: 1;
+    margin-bottom: 20px;
+}
+
+.btn-box {
+    display: flex;
+    gap: 10px;
+    flex-wrap: wrap;
+    padding: 20px 0;
+    border-top: 1px solid #ebeef5;
+}
+
+.btn-box .el-button {
+    min-width: 80px;
+}
+
+.form-box {
+    padding: 20px;
+}
+
+.row-box {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+}
+
+.row-box .el-input {
+    flex: 1;
+}
+
+.dialog-footer {
+    display: flex;
+    justify-content: flex-end;
+    gap: 10px;
+}
+
+/* 响应式设计 */
+@media (max-width: 768px) {
+    .content {
+        padding: 10px;
+    }
+    
+    .btn-box {
+        flex-direction: column;
+    }
+    
+    .row-box {
+        flex-direction: column;
+        align-items: stretch;
+        gap: 5px;
+    }
+}
